#51 버그 수정

+++ app/src/main/java/com/mashup/nawainvitation/data/model/response/InvitationsResponse.kt
@@ -23,8 +23,8 @@ data class InvitationsResponse(
         @SerializedName("invitationRoadAddressName")
         val invitationRoadAddressName: String?,
         @SerializedName("x")
-        val x: Int?,
+        val x: Double?,
         @SerializedName("y")
-        val y: Int?
+        val y: Double?
     )
 }
